requirements/base.txt
Django==5.0.4
djangorestframework==3.15.2
celery==5.4.0
redis==5.0.4
psycopg2-binary==2.9.9
django-environ==0.11.2
django-filter==24.2
djangorestframework-simplejwt==5.3.1
